fix hilt and user repo testing
