+++ app/src/main/java/ch/epfl/sdp/blindly/main_screen/match/my_matches/MyMatchesFragment.kt
@@ -89,7 +89,7 @@ class MyMatchesFragment : Fragment(), MyMatchesAdapter.OnItemClickListener {
      *
      * @param myMatches List of user's matches
      */
-    private suspend fun setAdapterOnMainThread(myMatches: ArrayList<MyMatch>) {
+    private suspend fun setAdapterOnMainThread(myMatches: MutableList<MyMatch>) {
         withContext(Dispatchers.Main) {
             setupRecylerView(myMatches)
         }
@@ -100,7 +100,7 @@ class MyMatchesFragment : Fragment(), MyMatchesAdapter.OnItemClickListener {
      *
      * @param myMatches List of user's matches
      */
-    private fun setupRecylerView(myMatches: ArrayList<MyMatch>) {
+    private fun setupRecylerView(myMatches: MutableList<MyMatch>) {
         myMatchesRecyclerView = fragView.findViewById(R.id.my_matches_recyler_view)
         myMatchesRecyclerView.layoutManager = LinearLayoutManager(context)
         adapter = MyMatchesAdapter(myMatches, arrayListOf(), requireContext(), this)
